paulolima/pratica03/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

openai==0.27.2
pyodbc==4.0.39
python-dotenv==0.21.0
